move job ID to the job module
