+++ src/lib.rs
@@ -5,6 +5,7 @@ mod cli;
 mod coordinator;
 mod fake_runner;
 mod glue;
+mod job;
 mod runner;
 
 use clap::Parser;
